+++ lib/vagrant/action/builtin/box_add.rb
@@ -476,6 +476,8 @@ module Vagrant
               # Actually not sure what causes this, but its always
               # in a case that isn't true.
               return false
+            rescue Errno::EISDIR
+              return false
             rescue Errno::ENOENT
               return false
             end