+++ CHANGELOG.md
@@ -22,6 +22,7 @@ BUG FIXES:
   - providers/docker: default proxy VM won't use HGFS [GH-3687]
   - providers/docker: fix container linking [GH-3719]
   - providers/docker: Port settings expose to host properly. [GH-3723]
+  - provisioners/puppet: Separate module paths with ';' on Windows. [GH-3731]
 
 ## 1.6.1 (May 7, 2014)